pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dumpcode"
version = "1.2.0"
description = "Codebase dumper with auto-versioning, dynamic profiles, and OSC52 clipboard support."
requires-python = ">=3.9"
authors = [
    {name = "Florent Lejoly", email = "florent.lejoly@gmail.com"}
]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "pathspec>=0.11.0",
]

[project.optional-dependencies]
token-counting = ["tiktoken>=0.5.0"]
dev = ["pytest", "pytest-cov", "ruff", "mypy"]

# NEW: AI provider dependencies
claude = ["anthropic>=0.40.0"]
gemini = ["google-generativeai>=0.8.0"]
openai = ["openai>=1.50.0"]
deepseek = ["openai>=1.50.0"]  # DeepSeek uses OpenAI-compatible API

# Convenience: install all AI providers
ai = [
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
]

# Full installation
all = [
    "tiktoken>=0.5.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dumpcode = "dumpcode.main:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]
exclude = [
    "tests*",
    "*.dump_config.json",
    "*.egg-info*",
    ".github*",
    ".git*",
]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
markers = [
    "edge_case: marks tests as edge case tests (deselect with '-m \"not edge_case\"')",
]
